:bug: :: fix circular reference
