+++ src/main/java/com/sam/shallwego/domain/member/service/MemberService.java
@@ -7,8 +7,6 @@ import com.sam.shallwego.domain.member.ro.LoginRO;
 import com.sam.shallwego.domain.member.ro.SignRO;
 import com.sam.shallwego.global.jwt.JwtUtil;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.ReactiveUserDetailsService;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,20 +16,11 @@ import reactor.core.scheduler.Schedulers;
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class MemberService implements ReactiveUserDetailsService {
+public class MemberService {
 
     private final PasswordEncoder passwordEncoder;
     private final MemberRepository memberRepository;
     private final JwtUtil jwtUtil;
-
-    @Override
-    @Transactional(readOnly = true)
-    public Mono<UserDetails> findByUsername(String username) {
-        return Mono.fromCallable(() -> memberRepository.findByUsername(username)
-                .orElseThrow(Member.NotExistsException::new))
-                .cast(UserDetails.class)
-                .subscribeOn(Schedulers.boundedElastic());
-    }
 
     public Mono<SignRO> registerMember(final SignDto signDto) {
         return Mono.fromCallable(() -> memberRepository.save(signDto.toEntity(passwordEncoder)))
